💄 SignupPage => RegisterPage로 이름 수정 - Signup, RegisterFormBox 파일에 작성하지 않고 index파일에 모든 코드를 넣는걸로 수정 - dispatch는 모두 커스텀 hooks로 관리하도록 바꿈 - LoginSignup.scss 폰트 파일 경로 http:// schem 추가
