--- src/pages/Signup/RegisterFormBox/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './RegisterFormBox';
