--- src/pages/Signup/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './Signup';
